Fix all comments link for guests
The jquery selector was depending on the comment form to be there.
Logged out users don't have a comment form and therefore, clicking the
"All Comments" link resulted all comments being removed from the DOM.

The fix adds a .comments div inside .Comments-section so that there is a
proper section containing all the comments.

https://github.com/BiteKollektiv/fairnopoly/issues/42